feat: operator hold by partition facet
