+++ packages/ats/contracts/Configuration.ts
@@ -126,7 +126,7 @@ export const CONTRACT_NAMES = [
   "ProtectedPartitionsFacet",
   "HoldFacet",
   "HoldByPartitionFacet",
-  "HoldManagementFacet",
+  "OperatorHoldByPartitionFacet",
   "TimeTravel",
   "KycFacet",
   "SsiManagementFacet",
