+++ packages/ats/contracts/contracts/constants/resolverKeys.sol
@@ -234,6 +234,9 @@ bytes32 constant _HOLD_TOKEN_HOLDER_SUSTAINABILITY_PERFORMANCE_TARGET_RATE_RESOL
 // keccak256("security.token.standard.hold.management.resolverKey")
 bytes32 constant _HOLD_MANAGEMENT_RESOLVER_KEY = 0xaab5a0e0978ad146ca8dc61d16bab0212224eadf68bd08e3c66600ee4f59c12a;
 
+// keccak256("security.token.standard.operator.hold.by.partition.resolverKey")
+bytes32 constant _OPERATOR_HOLD_BY_PARTITION_RESOLVER_KEY = 0x985830ee03940f1c26f2142776f45b16bdbeba9ae580393ab4e7fea7ade58a32;
+
 // keccak256('security.token.standard.controller.hold.by.partition.resolverKey');
 bytes32 constant _CONTROLLER_HOLD_BY_PARTITION_RESOLVER_KEY = 0x9e49506d2dfd484ed2aa6f2fd6f90a9efd8ae79466f93fa70571a95ddda4659c;
