--- packages/ats/contracts/contracts/facets/layer_1/hold/HoldManagementFacet.sol
@@ -1,26 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-pragma solidity >=0.8.0 <0.9.0;
-
-import { IHoldManagement } from "./IHoldManagement.sol";
-import { HoldManagement } from "./HoldManagement.sol";
-import { IStaticFunctionSelectors } from "../../../infrastructure/proxy/IStaticFunctionSelectors.sol";
-import { _HOLD_MANAGEMENT_RESOLVER_KEY } from "../../../constants/resolverKeys.sol";
-
-contract HoldManagementFacet is HoldManagement, IStaticFunctionSelectors {
-    function getStaticResolverKey() external pure override returns (bytes32 staticResolverKey_) {
-        staticResolverKey_ = _HOLD_MANAGEMENT_RESOLVER_KEY;
-    }
-
-    function getStaticFunctionSelectors() external pure override returns (bytes4[] memory staticFunctionSelectors_) {
-        uint256 selectorIndex = 1;
-        staticFunctionSelectors_ = new bytes4[](selectorIndex);
-        unchecked {
-            staticFunctionSelectors_[--selectorIndex] = this.operatorCreateHoldByPartition.selector;
-        }
-    }
-
-    function getStaticInterfaceIds() external pure override returns (bytes4[] memory staticInterfaceIds_) {
-        staticInterfaceIds_ = new bytes4[](1);
-        staticInterfaceIds_[0] = type(IHoldManagement).interfaceId;
-    }
-}
