+++ packages/ats/contracts/contracts/facets/layer_1/hold/IHold.sol
@@ -2,10 +2,9 @@
 pragma solidity >=0.8.0 <0.9.0;
 
 import { IHoldTypes } from "./IHoldTypes.sol";
-import { IHoldManagement } from "./IHoldManagement.sol";
 
-/// @title IHold — off-chain umbrella for the Hold Management domain
-/// @notice Aggregates IHoldManagement into a TypeChain-friendly handle for tests and SDK consumers.
+/// @title IHold — off-chain umbrella for the Hold domain
+/// @notice Aggregates hold types into a TypeChain-friendly handle for tests and SDK consumers.
 /// @dev DO NOT inherit this interface from any facet contract. Each facet must `is` exactly its
 ///      per-facet interface. This umbrella contains no function declarations of its own.
-interface IHold is IHoldTypes, IHoldManagement {}
+interface IHold is IHoldTypes {}
