--- packages/ats/contracts/contracts/test/testTimeTravel/facetsTimeTravel/holdManagement/HoldManagementFacetTimeTravel.sol
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-pragma solidity >=0.8.0 <0.9.0;
-
-import { TimeTravelProvider } from "../../timeTravel/TimeTravelProvider.sol";
-import { HoldManagementFacet } from "../../../../facets/layer_1/hold/HoldManagementFacet.sol";
-
-contract HoldManagementFacetTimeTravel is HoldManagementFacet, TimeTravelProvider {
-    // solhint-disable-next-line no-empty-blocks
-}
