+++ packages/ats/contracts/scripts/domain/bondFixedRate/createConfiguration.ts
@@ -106,7 +106,7 @@ const BOND_FIXED_RATE_FACETS = [
   "ClearingRedeemFacet",
   "ClearingTransferFacet",
   "HoldFacet",
-  "HoldManagementFacet",
+  "OperatorHoldByPartitionFacet",
   "ControllerHoldByPartitionFacet",
   "ControllerByPartitionFacet",
   "ProtectedHoldByPartitionFacet",
