+++ packages/ats/contracts/scripts/domain/orchestratorLibraries.ts
@@ -121,7 +121,7 @@ export const LIBRARY_DEPENDENT_FACETS: Record<string, Array<keyof typeof LIBRARY
   BurnFacet: ["tokenCoreOps"],
   AdjustBalancesFacet: ["tokenCoreOps"],
   // HoldOps dependencies - hold/lock operations
-  HoldManagementFacet: ["holdOps"],
+  OperatorHoldByPartitionFacet: ["holdOps"],
   HoldFacet: ["holdOps"],
   HoldByPartitionFacet: ["holdOps"],
   // ClearingOps dependencies - clearing transfer operations
